requirements.txt
# factorytx Requirements
Flask
Jinja2
MarkupSafe
pyyaml
Werkzeug
argh
boto
cffi
click
cryptography
enum34
idna
ipaddress
itsdangerous
jsonschema
pathtools
pyasn1
pycparser
pymongo
requests
rfc5424syslog
semantic-version
simpleeval
six
watchdog
SQLAlchemy
# Plugin Requirements
freeopcua
ftputil
futures
numpy
pandas
path.py
psycopg2
pysmb
python-dateutil
pytz
trollius
ujson
xlrd
# Testing Requirements
pytest
fixtures
pylint
pyflake
pep8
dateparser
cov-core
coverage
mock
nose-cov
testfixtures
